The Basal-like and Classical gene list used in Tiriac et al 2018.

The Basal-like and Classical gene list used in Tiriac et al 2018.

data(Tiriac_2018)

Format

a data.frame with the following columns:

SYMBOL a character string representing the symbol of the gene.

Class a character string representing the class (Basal-like or Classical) associated to the gene.

References

Tiriac H, Belleau P, Engle DD, Plenker D, Deschênes A, Somerville TDD, Froeling FEM, Burkhart RA, Denroche RE, Jang GH, Miyabayashi K, Young CM, Patel H, Ma M, LaComb JF, Palmaira RLD, Javed AA, Huynh JC, Johnson M, Arora K, Robine N, Shah M, Sanghvi R, Goetz AB, Lowder CY, Martello L, Driehuis E, LeComte N, Askan G, Iacobuzio-Donahue CA, Clevers H, Wood LD, Hruban RH, Thompson E, Aguirre AJ, Wolpin BM, Sasson A, Kim J, Wu M, Bucobo JC, Allen P, Sejpal DV, Nealon W, Sullivan JD, Winter JM, Gimotty PA, Grem JL, DiMaio DJ, Buscaglia JM, Grandgenett PM, Brody JR, Hollingsworth MA, O'Kane GM, Notta F, Kim E, Crawford JM, Devoe C, Ocean A, Wolfgang CL, Yu KH, Li E, Vakoc CR, Hubert B, Fischer SE, Wilson JM, Moffitt R, Knox J, Krasnitz A, Gallinger S, Tuveson DA. Organoid Profiling Identifies Common Responders to Chemotherapy in Pancreatic Cancer. Cancer Discov. 2018 Sep;8(9):1112-1129. doi: 10.1158/2159-8290.CD-18-0349. Epub 2018 May 31. PMID: 29853643; PMCID: PMC6125219.

See also

Examples

## Load Tiriac et al 2018 gene list used for 
## Basal-like/Classical classification
data(Tiriac_2018)

## The dataset contains 2 columns
## The first column has the gene symbols
## The second column the class assignation (Basal-like or Classical)
head(Tiriac_2018)
#>     SYMBOL     Class
#> 1    MUC13 Classical
#> 2    GPR35 Classical
#> 3    USH1C Classical
#> 4 BAIAP2L2 Classical
#> 5  GAL3ST1 Classical
#> 6     VIL1 Classical